# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Veil API
  description: REST API for disposable email aliases, virtual cards, and burner phone numbers
  version: 0.1.0
  contact:
    name: Veil Team

servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.veil.sh
    description: Production

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
          required: [code, message]
      required: [error]

    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string, nullable: true }
        plan_tier: { type: string, enum: [free, basic, pro] }

    AuthResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        access_token: { type: string }
        refresh_token: { type: string }

    EmailAlias:
      type: object
      properties:
        id: { type: string, format: uuid }
        alias: { type: string, example: "abc12345@veilmail.io" }
        display_name: { type: string, nullable: true }
        forwarding_address: { type: string, format: email }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }

    PhoneNumber:
      type: object
      properties:
        id: { type: string, format: uuid }
        phone_number: { type: string, example: "+15551234567" }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }

    VirtualCard:
      type: object
      properties:
        id: { type: string, format: uuid }
        last_four: { type: string, example: "4242" }
        brand: { type: string, example: "visa" }
        status: { type: string, enum: [active, frozen, canceled] }
        spending_limit_amount: { type: integer, nullable: true }
        spending_limit_interval: { type: string, nullable: true }

    Subscription:
      type: object
      properties:
        plan_tier: { type: string, enum: [free, basic, pro] }
        status: { type: string, enum: [active, canceled, past_due] }
        current_period_end: { type: string, format: date-time, nullable: true }

paths:
  /api/health:
    get:
      summary: Health check
      tags: [Health]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  service: { type: string }
                  timestamp: { type: string, format: date-time }

  /api/auth/signup:
    post:
      summary: Create a new account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                display_name: { type: string }
      responses:
        '201': { description: Account created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '409': { description: Email already exists, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /api/auth/login:
    post:
      summary: Login to existing account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Login successful, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  /api/aliases:
    get:
      summary: List email aliases
      tags: [Aliases]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of aliases
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/EmailAlias' }
    post:
      summary: Create a new email alias
      tags: [Aliases]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [forwarding_address]
              properties:
                forwarding_address: { type: string, format: email }
                display_name: { type: string }
      responses:
        '201': { description: Alias created, content: { application/json: { schema: { type: object, properties: { data: { $ref: '#/components/schemas/EmailAlias' } } } } } }

  /api/aliases/{id}:
    delete:
      summary: Delete an email alias
      tags: [Aliases]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Alias deleted }

  /api/numbers:
    get:
      summary: List phone numbers
      tags: [Numbers]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PhoneNumber' }
    post:
      summary: Rent a phone number
      tags: [Numbers]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                area_code: { type: string, example: "415" }
      responses:
        '201': { description: Number rented }

  /api/cards:
    get:
      summary: List virtual cards
      tags: [Cards]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/VirtualCard' }
    post:
      summary: Create a virtual card
      tags: [Cards]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                spending_limit_amount: { type: integer, example: 5000 }
                spending_limit_interval: { type: string, enum: [daily, weekly, monthly, yearly] }
      responses:
        '201': { description: Card created }

  /api/subscriptions:
    get:
      summary: Get current subscription
      tags: [Subscriptions]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current subscription details }

  /api/subscriptions/checkout:
    post:
      summary: Create Stripe checkout session
      tags: [Subscriptions]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [tier]
              properties:
                tier: { type: string, enum: [basic, pro] }
      responses:
        '200': { description: Checkout URL }